serverless/cloud-run-function/requirements.txt
functions-framework==3.*
google-auth-oauthlib
google-api-python-client
google-auth
psycopg2
google-auth-httplib2
psycopg2-binary
